python/file-exporter/requirements.txt
prometheus_client==0.7.0
pyyaml==3.12
